refactor: altera a arquitetura implementando domains e corrige imports
